+++ src/v5/codec/encode.rs
@@ -311,7 +311,7 @@ mod tests {
         let mut v = BytesMut::with_capacity(1024);
         packet.encode(&mut v, packet.encoded_size(1024) as u32).unwrap();
         assert_eq!(expected.len(), v.len());
-        assert_eq!(&expected[..], &v[..]);
+        assert_eq!(expected, &v[..]);
     }
 
     #[test]